requirements (3).txt
pandas==2.2.3
streamlit==1.39.0
numpy==2.1.1
pillow==10.4.0
requests==2.32.3